refaktorert kode og endret navn fra ettersendingDb til ettersending
