+++ src/main/kotlin/no/nav/familie/ef/mottak/repository/DokumentasjonsbehovRepository.kt
@@ -3,7 +3,6 @@ package no.nav.familie.ef.mottak.repository
 import no.nav.familie.ef.mottak.repository.domain.Dokumentasjonsbehov
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
-import java.util.*
 
 @Repository
 interface DokumentasjonsbehovRepository : JpaRepository<Dokumentasjonsbehov, String>
